// src/app/api-client/model/pagedRequestTaskFilterDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskFilterDto } from './taskFilterDto';


export interface PagedRequestTaskFilterDto { 
    size?: number;
    page?: number;
    sort?: PagedRequestTaskFilterDto.SortEnum;
    filter?: TaskFilterDto;
}
export namespace PagedRequestTaskFilterDto {
    export const SortEnum = {
        Asc: 'ASC',
        Desc: 'DESC'
    } as const;
    export type SortEnum = typeof SortEnum[keyof typeof SortEnum];
}