// src/app/api-client/model/pagedResultTaskDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskDto } from './taskDto';


export interface PagedResultTaskDto { 
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    last?: boolean;
    entities?: Array<TaskDto>;
}