// src/app/api-client/model/taskFilterDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TaskFilterDto { 
    processDefinitionKey?: string;
    taskId?: string;
    assigneeId?: string;
    groupIds?: Array<string>;
    userId?: string;
    businessKey?: string;
    taskName?: string;
    startDate?: string;
    endDate?: string;
}